requirements.txt
ansible==9.3.0
ansible-core==2.16.4
cffi==1.16.0
cryptography==42.0.5
Jinja2==3.1.3
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.21
PyYAML==6.0.1
resolvelib==1.0.1
